requirements.txt
numpy
falcon
falcon-cors
celery[redis]
redis
jsonpickle
scikit-learn
python-dotenv

# gunicorn
waitress # for dev purposes
# flower for later